+++ crates/pi_scene_shell/src/lighting_shadow.rs
@@ -3,10 +3,10 @@ use pi_bevy_render_plugin::constant::texture_sampler::{ColorFormat, DepthStencil
 
 #[derive(Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord)]
 pub struct LightLimitInfo {
-    pub max_direct_light_count: u32,
-    pub max_point_light_count: u32,
-    pub max_spot_light_count: u32,
-    pub max_hemi_light_count: u32,
+    pub max_direct_light_count: u16,
+    pub max_point_light_count: u16,
+    pub max_spot_light_count: u16,
+    pub max_hemi_light_count: u16,
 }
 
 #[derive(Resource)]
@@ -17,7 +17,7 @@ pub struct SceneLightLimit(pub LightLimitInfo);
 
 #[derive(Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord)]
 pub struct ShadowLimitInfo {
-    pub max_count: u32,
+    pub max_count: u16,
     pub max_width: u32,
     pub max_height: u32,
     pub color_format: ColorFormat,
